Changed filter popup to UIMenu
