--- Zotero/Scenes/Detail/Items/ItemsFilterCoordinator.swift
@@ -1,79 +1,0 @@
-//
-//  ItemsFilterCoordinator.swift
-//  Zotero
-//
-//  Created by Michal Rentka on 22.03.2023.
-//  Copyright © 2023 Corporation for Digital Scholarship. All rights reserved.
-//
-
-import UIKit
-
-protocol ItemsFilterCoordinatorDelegate: AnyObject {
-    func showTagPicker(libraryId: LibraryIdentifier, selected: Set<String>, picked: @escaping ([Tag]) -> Void)
-}
-
-protocol FiltersDelegate: AnyObject {
-    var currentLibrary: Library { get }
-
-    func downloadsFilterDidChange(enabled: Bool)
-    func tagSelectionDidChange(selected: Set<String>)
-    func tagOptionsDidChange()
-}
-
-final class ItemsFilterCoordinator: NSObject, Coordinator {
-    weak var parentCoordinator: Coordinator?
-    var childCoordinators: [Coordinator]
-    weak var navigationController: UINavigationController?
-
-    private let filters: [ItemsFilter]
-    private unowned let mainCoordinatorDelegate: MainCoordinatorDelegate
-    private unowned let controllers: Controllers
-    private weak var filtersDelegate: BaseItemsViewController?
-
-    init(
-        filters: [ItemsFilter],
-        filtersDelegate: BaseItemsViewController,
-        navigationController: NavigationViewController,
-        mainCoordinatorDelegate: MainCoordinatorDelegate,
-        controllers: Controllers
-    ) {
-        self.filters = filters
-        self.navigationController = navigationController
-        self.mainCoordinatorDelegate = mainCoordinatorDelegate
-        self.controllers = controllers
-        self.filtersDelegate = filtersDelegate
-        childCoordinators = []
-
-        super.init()
-
-        navigationController.dismissHandler = { [weak self] in
-            guard let self = self else { return }
-            self.parentCoordinator?.childDidFinish(self)
-        }
-    }
-
-    func start(animated: Bool) {
-        guard let viewModel = mainCoordinatorDelegate.sharedTagFilterViewModel else { return }
-        let tagController = TagFilterViewController(viewModel: viewModel)
-        tagController.view.translatesAutoresizingMaskIntoConstraints = false
-        tagController.delegate = filtersDelegate
-        filtersDelegate?.tagFilterDelegate = tagController
-
-        let downloadsFilterEnabled = filters.contains(where: { $0.isDownloadedFilesFilter })
-        let controller = ItemsFilterViewController(downloadsFilterEnabled: downloadsFilterEnabled, tagFilterController: tagController)
-        controller.delegate = filtersDelegate
-        controller.coordinatorDelegate = self
-        navigationController?.setViewControllers([controller], animated: animated)
-    }
-}
-
-extension ItemsFilterCoordinator: ItemsFilterCoordinatorDelegate {
-    func showTagPicker(libraryId: LibraryIdentifier, selected: Set<String>, picked: @escaping ([Tag]) -> Void) {
-        guard let dbStorage = controllers.userControllers?.dbStorage else { return }
-        let state = TagPickerState(libraryId: libraryId, selectedTags: selected)
-        let handler = TagPickerActionHandler(dbStorage: dbStorage)
-        let viewModel = ViewModel(initialState: state, handler: handler)
-        let controller = TagPickerViewController(viewModel: viewModel, saveAction: picked)
-        navigationController?.pushViewController(controller, animated: true)
-    }
-}
