+++ Zotero/Scenes/Detail/Trash/Views/TrashViewController.swift
@@ -151,7 +151,8 @@ final class TrashViewController: BaseItemsViewController {
 
     private func process(action: ItemAction.Kind, for selectedKeys: Set<TrashKey>, button: UIBarButtonItem?, completionAction: ((Bool) -> Void)?) {
         switch action {
-        case .createParent, .retrieveMetadata, .duplicate, .trash, .copyBibliography, .copyCitation, .share, .addToCollection, .removeFromCollection, .removeFromRecentlyRead:
+        case .createParent, .retrieveMetadata, .duplicate, .trash, .copyBibliography, .copyCitation, .share, .addToCollection, .removeFromCollection, .removeFromRecentlyRead,
+             .filter, .sort, .debugReader:
             // These actions are not available in trash collection
             break
 
@@ -172,21 +173,11 @@ final class TrashViewController: BaseItemsViewController {
             viewModel.process(action: .restoreItems(selectedKeys))
             completionAction?(true)
 
-        case .filter:
-            guard let button else { return }
-            coordinatorDelegate?.showFilters(filters: viewModel.state.filters, filtersDelegate: self, button: button)
-
-        case .sort:
-            break
-
         case .download:
             viewModel.process(action: .download(selectedKeys))
 
         case .removeDownload:
             viewModel.process(action: .removeDownloads(selectedKeys))
-
-        case .debugReader:
-            break
         }
     }
 
@@ -340,6 +331,14 @@ extension TrashViewController: ItemsToolbarControllerDelegate {
 
     func sortTypeChanged(_ sortType: ItemsSortType) {
         viewModel.process(action: .setSortType(sortType))
+    }
+
+    func downloadsFilterChanged(enabled: Bool) {
+        if enabled {
+            viewModel.process(action: .enableFilter(.downloadedFiles))
+        } else {
+            viewModel.process(action: .disableFilter(.downloadedFiles))
+        }
     }
 }
